--- src/main/java/de/hamburgchimps/apple/notes/liberator/data/EmbeddedObjectDataType.java
@@ -1,6 +1,0 @@
-package de.hamburgchimps.apple.notes.liberator.data;
-
-public enum EmbeddedObjectDataType {
-    TABLE,
-    FILE;
-}
