+++ src/main/java/de/hamburgchimps/apple/notes/liberator/data/Markdownable.java
@@ -1,4 +1,4 @@
-package de.hamburgchimps.apple.notes.liberator.data.format;
+package de.hamburgchimps.apple.notes.liberator.data;
 
 public interface Markdownable {
     String toMarkdown();
