+++ src/main/java/de/hamburgchimps/apple/notes/liberator/data/embedded/EmbeddedObjectData.java
@@ -1,4 +1,4 @@
-package de.hamburgchimps.apple.notes.liberator.data;
+package de.hamburgchimps.apple.notes.liberator.data.embedded;
 
 public interface EmbeddedObjectData {
     EmbeddedObjectDataType getType();
