+++ src/main/java/de/hamburgchimps/apple/notes/liberator/data/embedded/EmbeddedObjectDataType.java
@@ -1,0 +1,6 @@
+package de.hamburgchimps.apple.notes.liberator.data.embedded;
+
+public enum EmbeddedObjectDataType {
+    TABLE,
+    FILE;
+}
