+++ src/main/java/de/hamburgchimps/apple/notes/liberator/data/format/Link.java
@@ -1,5 +1,7 @@
 package de.hamburgchimps.apple.notes.liberator.data.format;
 
+import de.hamburgchimps.apple.notes.liberator.data.Markdownable;
+
 public record Link(String text, String url) implements Markdownable {
     @Override
     public String toMarkdown() {
